#38: Auth Middleware Implementation (#44)
* feat(auth): auth init

* chore(log): remove debu log

* feat(auth): fix error handling and add .env

* feat(auth): validation on response object

* chore(comments): update comment and remove res

* feat(auth): verifyAllowedAccess init

* feat(logger): add logger to auth client

* feat(): verify allowed access implement and add env

* fix(dac): remove auth from dac endpoints

* fix(revert): rever dac changes

* feat(auth): grab study from req

* chore(auth): detailed comments

* feat(auth): update to new env variables

* feat(verifyUser): user token parser init

* fix(build): resolve build issue

* refactor(config): refactor auth config

* fix(config): remove previous config

* fix(auth): refactor auth middleware implemenation

* fix(auth): remove action form middleware

* feat(auth): create helper functions and apply async

* chore(debug): remove debug

* chore(imports): remove unused import and export token

* chore(debug): remove unused imports and comments

* feat(env): add auth env variables

* feat(auth): add hasAllowedAccess checks in studies

* chore(comment): update comments

* refactor(auth): move and rm retrieveUserTokenInformation

* chore(comments): update comments for helper functions

* fix(.env): remove unused .env var

* fix(error): update error response for retrieveUserTokenInformation

* fix(env): ADB_HOST back to DB_HOST

* refactor(auth): rename auth client

* chore: remove extra char

* refactor(auth): remove redundant function

* chore: unused import

* feat(auth): add pcglUserSessionResult type

* feat(dac): add auth to dac endpoints

* feat(auth): create helper function to extract user studies

* feat(auth): admin lyric middleware

* development setup documentation

* update development setup doc

* fix(auth): adjust to use study_curator array

* feat(auth): update types

* feat(zod): update zod validation for user response

* fix(auth): update comment and add admin check to getAllStudies

* chore(auth): update comment

* feat(env): readd missing env and organize

* fix(auth): rename to admin middleware and have enabled check

* fix(swagger): resolve auth token not working in swagger

* fix(url): safer url join

* chore(file): remove doc

* fix(auth): remove unused env and update docs

* feat(config): update env auth names for clarity

* fix(typo): add period

* feat(auth): create authz client

* fix(auth): admin checks only for mutation endpoints

* chore(comments): update errors and add note

* fix(error): update errors types and handling

---------

+++ src/controllers/dacController.ts
@@ -78,6 +78,11 @@ const createDac = validateRequest(createDacData, async (req, res, next) => {
 		const dacSvc = await dacService(database);
 
 		const dacFields = req.body;
+		const user = req.user;
+
+		if (!user?.isAdmin) {
+			throw new lyricProvider.utils.errors.Forbidden('Must be an admin user.');
+		}
 
 		const result = await dacSvc.saveDac(dacFields);
 
@@ -94,6 +99,11 @@ const deleteDac = validateRequest(deleteDacByIdData, async (req, res, next) => {
 		const dacSvc = await dacService(database);
 
 		const dacId = req.params.dacId;
+		const user = req.user;
+
+		if (!user?.isAdmin) {
+			throw new lyricProvider.utils.errors.Forbidden('Must be an admin user.');
+		}
 
 		const result = await dacSvc.deleteDacById(dacId);
 
@@ -117,6 +127,11 @@ const updateDac = validateRequest(updateDacByIdData, async (req, res, next) => {
 
 		const dacId = req.params.dacId;
 		const dacFields = req.body;
+		const user = req.user;
+
+		if (!user?.isAdmin) {
+			throw new lyricProvider.utils.errors.Forbidden('Must be an admin user.');
+		}
 
 		const result = await dacSvc.updateDacById(dacId, dacFields);
